refactor(request): refactor CRUD requests
