+++ jina/types/request/delete.py
@@ -1,0 +1,5 @@
+from . import Request
+
+
+class DeleteRequest(Request):
+    pass
