+++ jina/types/request/index.py
@@ -1,0 +1,6 @@
+from . import Request
+from .mixin import *
+
+
+class IndexRequest(Request, DocsPropertyMixin, GroundtruthPropertyMixin):
+    pass
