+++ jina/types/request/train.py
@@ -1,0 +1,6 @@
+from . import Request
+from .mixin import *
+
+
+class TrainRequest(Request, DocsPropertyMixin, GroundtruthPropertyMixin):
+    pass
