+++ jina/types/request/update.py
@@ -1,0 +1,6 @@
+from . import Request
+from .mixin import *
+
+
+class SearchRequest(Request, DocsPropertyMixin, GroundtruthPropertyMixin):
+    pass
